Graphics converted over to using coordinates in {0, 1}. Switched over to a custom Event structure (rather than SDL_Event) which gives access to mouse coords in {0, 1} instead of just pixel coordinates. Graphics are distorted for now -- need to rework their on-screen sizes.

+++ include/base.hpp
@@ -16,7 +16,7 @@
 typedef unsigned short CardID;
 
 typedef unsigned short LocationID;
-typedef avl::AABB2<int> Location;
+typedef avl::AABB2<double> Location;
 typedef std::map<LocationID, Location> LocationMap;
